# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Posts API",
        "version": "0.1"
    },
    "paths": {
        "/posts/{$id}": {
            "get": {
                "description": "Get a posts",
                "responses": {
                    "200": {
                        "description": "Get a posts successfully !"
                    }
                }
            },
            "put": {
                "description": "Update a posts",
                "requestBody": {
                    "description": "Data of the new posts update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author_id": {
                                        "type": "number"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "tag": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "category_id": {
                                        "type": "string"
                                    },
                                    "updated_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update posts successfully !"
                    },
                    "401": {
                        "description": "There was an error during the update process !"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "Get a list of posts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get posts successfully !"
                    }
                }
            },
            "post": {
                "description": "Create a new posts",
                "requestBody": {
                    "description": "Data of the new posts",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author_id": {
                                        "type": "number"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "tag": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "category_id": {
                                        "type": "string"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create posts successfully !"
                    },
                    "401": {
                        "description": "There was an error during creation !"
                    }
                }
            }
        },
        "/posts/{$id}/like": {
            "post": {
                "description": "Like a posts",
                "requestBody": {
                    "description": "Data of the like posts",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "author_id": {
                                        "type": "number"
                                    },
                                    "posts_id": {
                                        "type": "number"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "You have liked this posts !"
                    },
                    "200": {
                        "description": "You have liked this posts !"
                    }
                }
            }
        }
    }
}